+++ packages/application/test/radix.test.ts
@@ -70,7 +70,7 @@ import {
 	Wallet,
 	WalletT,
 } from '../src'
-import { Amount, AmountT, NetworkT } from '@radixdlt/primitives'
+import { Amount, AmountT, Network } from '@radixdlt/primitives'
 
 import {
 	log,
@@ -88,7 +88,7 @@ import {
 	IntendedAction,
 	SimpleExecutedTransaction,
 	TransactionIntent,
-} from '..'
+} from '../'
 import { signatureFromHexStrings } from '@radixdlt/crypto/test/utils'
 import { makeSigningKeyChainWithFunds } from '@radixdlt/account/test/utils'
 import { UInt256 } from '@radixdlt/uint256'
@@ -97,7 +97,7 @@ import { createWallet } from './util'
 export const makeWalletWithFunds = (): WalletT => {
 	return Wallet.create({
 		signingKeychain: makeSigningKeyChainWithFunds(),
-		network: NetworkT.BETANET,
+		network: Network.MAINNET,
 	})
 }
 
@@ -355,7 +355,7 @@ describe('radix_high_level_api', () => {
 
 		radix.activeAddress.subscribe(
 			address => {
-				expect(address.network).toBe(NetworkT.BETANET)
+				expect(address.network).toBe(Network.MAINNET)
 				done()
 			},
 			error => done(error),
@@ -734,7 +734,7 @@ describe('radix_high_level_api', () => {
 		subs.add(
 			radix.__wallet.subscribe(_w => {
 				const expectedValues = [
-						{ pkIndex: 0, tokenBalancesCount: 1 },
+					{ pkIndex: 0, tokenBalancesCount: 1 },
 					{ pkIndex: 1, tokenBalancesCount: 4 },
 					{ pkIndex: 2, tokenBalancesCount: 1 },
 					{ pkIndex: 3, tokenBalancesCount: 1 },
@@ -918,7 +918,7 @@ describe('radix_high_level_api', () => {
 		radix.login(keystoreForTest.password, loadKeystore)
 
 		const mockedValidatorAddr = ValidatorAddress.fromUnsafe(
-			'vb1qvx0emaq0tua6md7wu9c047mm5krrwnlfl8c7ws3jm2s9uf4vxcyvrwrazy',
+			'vr1qvz3anvawgvm7pwvjs7xmjg48dvndczkgnufh475k2tqa2vm5c6cq4eae87',
 		)._unsafeUnwrap()
 
 		subs.add(
@@ -931,7 +931,7 @@ describe('radix_high_level_api', () => {
 						).toBe(true)
 						expect(
 							validator.ownerAddress.toString().slice(-4),
-						).toBe('9gwm')
+						).toBe('j7dt')
 						done()
 					})
 			}),
@@ -964,7 +964,7 @@ describe('radix_high_level_api', () => {
 		const transactionIntent = TransactionIntentBuilder.create()
 			.stakeTokens({
 				validator:
-					'vb1qvx0emaq0tua6md7wu9c047mm5krrwnlfl8c7ws3jm2s9uf4vxcyvrwrazy',
+					'vr1qvz3anvawgvm7pwvjs7xmjg48dvndczkgnufh475k2tqa2vm5c6cq4eae87',
 				amount: 10000,
 			})
 			.__syncBuildDoNotEncryptMessageIfAny(alice)
@@ -976,7 +976,7 @@ describe('radix_high_level_api', () => {
 				.subscribe(unsignedTx => {
 					expect(
 						(unsignedTx as { fee: AmountT }).fee.toString(),
-					).toEqual('7794')
+					).toEqual('21256')
 					done()
 				}),
 		)
@@ -1021,7 +1021,7 @@ describe('radix_high_level_api', () => {
 		const radix = Radix.create().__withAPI(mockedAPI)
 
 		subs.add(
-			radix.ledger.networkTransactionDemand().subscribe(result => {
+			radix.ledger.NetworkTransactionDemand().subscribe(result => {
 				expect(result.tps).toEqual(109)
 				done()
 			}),
@@ -1034,7 +1034,7 @@ describe('radix_high_level_api', () => {
 		const radix = Radix.create().__withAPI(mockedAPI)
 
 		subs.add(
-			radix.ledger.networkTransactionThroughput().subscribe(result => {
+			radix.ledger.NetworkTransactionThroughput().subscribe(result => {
 				expect(result.tps).toEqual(10)
 				done()
 			}),
@@ -1084,7 +1084,7 @@ describe('radix_high_level_api', () => {
 		radix.login(keystoreForTest.password, loadKeystore)
 
 		const expectedStakes = [
-			{ amount: 208, validator: '23', epochsUntil: 3 },
+			{ amount: 208, validator: '87', epochsUntil: 3 },
 		]
 		const expectedValues = [expectedStakes, expectedStakes] // should be unchanged between updates (deterministically mocked).
 		subs.add(
@@ -1145,7 +1145,7 @@ describe('radix_high_level_api', () => {
 					transferInput: {
 						to: bob,
 						amount: 1,
-						tokenIdentifier: 'xrd_rb1qya85pwq',
+						tokenIdentifier: 'xrd_tr1qyf0x76s',
 					},
 					userConfirmation: 'skip',
 					message: { plaintext, encrypt: false },
@@ -1183,7 +1183,7 @@ describe('radix_high_level_api', () => {
 		const bobPublicKey = bobPrivateKey.publicKey()
 		const bob = AccountAddress.fromPublicKeyAndNetwork({
 			publicKey: bobPublicKey,
-			network: NetworkT.BETANET,
+			network: Network.MAINNET,
 		})
 
 		const plaintext =
@@ -1214,7 +1214,7 @@ describe('radix_high_level_api', () => {
 			transferInput: {
 				to: bob,
 				amount: 1,
-				tokenIdentifier: 'xrd_rb1qya85pwq',
+				tokenIdentifier: 'xrd_tr1qyf0x76s',
 			},
 			userConfirmation: 'skip',
 			message: { plaintext, encrypt: true },
@@ -1305,12 +1305,12 @@ describe('radix_high_level_api', () => {
 		)._unsafeUnwrap()
 		const recipientAddress = AccountAddress.fromPublicKeyAndNetwork({
 			publicKey: recipientPK,
-			network: NetworkT.BETANET,
+			network: Network.MAINNET,
 		})
 		const tokenTransferInput: TransferTokensInput = {
 			to: recipientAddress,
 			amount: 1,
-			tokenIdentifier: 'xrd_rb1qya85pwq',
+			tokenIdentifier: 'xrd_tr1qyf0x76s',
 		}
 
 		const plaintext = 'Hey Bob, this is Alice.'
@@ -1451,7 +1451,7 @@ describe('radix_high_level_api', () => {
 		const tokenTransferInput: TransferTokensInput = {
 			to: bob,
 			amount: 1,
-			tokenIdentifier: 'xrd_rb1qya85pwq',
+			tokenIdentifier: 'xrd_tr1qyf0x76s',
 		}
 
 		let pollTXStatusTrigger: Observable<unknown>
@@ -1654,7 +1654,7 @@ describe('radix_high_level_api', () => {
 						stakeInput: {
 							amount: 1,
 							validator:
-								'vb1qvx0emaq0tua6md7wu9c047mm5krrwnlfl8c7ws3jm2s9uf4vxcyvrwrazy',
+								'vr1qvz3anvawgvm7pwvjs7xmjg48dvndczkgnufh475k2tqa2vm5c6cq4eae87',
 						},
 						userConfirmation: 'skip',
 						pollTXStatusTrigger: pollTXStatusTrigger,
@@ -1685,7 +1685,7 @@ describe('radix_high_level_api', () => {
 						unstakeInput: {
 							amount: 1,
 							validator:
-								'vb1qvx0emaq0tua6md7wu9c047mm5krrwnlfl8c7ws3jm2s9uf4vxcyvrwrazy',
+								'vr1qvz3anvawgvm7pwvjs7xmjg48dvndczkgnufh475k2tqa2vm5c6cq4eae87',
 						},
 						userConfirmation: 'skip',
 						pollTXStatusTrigger: pollTXStatusTrigger,
@@ -1781,7 +1781,7 @@ describe('radix_high_level_api', () => {
 		})
 	})
 
-	it('special signingKeychain with preallocated funds', done => {
+	it.skip('special signingKeychain with preallocated funds', done => {
 		const subs = new Subscription()
 
 		const walletWithFunds = makeWalletWithFunds()
@@ -1790,8 +1790,8 @@ describe('radix_high_level_api', () => {
 			.__withAPI(
 				of({
 					...mockRadixCoreAPI(),
-					networkId: (): Observable<NetworkT> => {
-						return of(NetworkT.BETANET).pipe(shareReplay(1))
+					networkId: (): Observable<Network> => {
+						return of(Network.MAINNET).pipe(shareReplay(1))
 					},
 				}),
 			)
@@ -1834,7 +1834,7 @@ describe('radix_high_level_api', () => {
 			const subs = new Subscription()
 
 			const xrdRRI = ResourceIdentifier.fromUnsafe(
-				'xrd_rb1qya85pwq',
+				'xrd_tr1qyf0x76s',
 			)._unsafeUnwrap()
 
 			const txID = TransactionIdentifier.create(
@@ -1859,7 +1859,7 @@ describe('radix_high_level_api', () => {
 			}
 
 			const wallet = makeWalletWithFunds()
-			const network = NetworkT.BETANET
+			const network = Network.MAINNET
 			const myAddress = AccountAddress.fromPublicKeyAndNetwork({
 				publicKey: wallet.__unsafeGetAccount().publicKey,
 				network,
@@ -1883,7 +1883,7 @@ describe('radix_high_level_api', () => {
 								type: ActionType.STAKE_TOKENS,
 								from: fromMe ? myAddress : bob,
 								validator: ValidatorAddress.fromUnsafe(
-									'vb1qgfqnj34dn7qp9wvf4l6rhw6hu3l82rcqh3rjtk080t75t888u98wkh3gjq',
+									'vr1qvz3anvawgvm7pwvjs7xmjg48dvndczkgnufh475k2tqa2vm5c6cq4eae87',
 								)._unsafeUnwrap(),
 								amount: Amount.fromUnsafe(1)._unsafeUnwrap(),
 							},
@@ -1891,7 +1891,7 @@ describe('radix_high_level_api', () => {
 								type: ActionType.UNSTAKE_TOKENS,
 								from: fromMe ? myAddress : bob,
 								validator: ValidatorAddress.fromUnsafe(
-									'vb1qgfqnj34dn7qp9wvf4l6rhw6hu3l82rcqh3rjtk080t75t888u98wkh3gjq',
+									'vr1qvz3anvawgvm7pwvjs7xmjg48dvndczkgnufh475k2tqa2vm5c6cq4eae87',
 								)._unsafeUnwrap(),
 								amount: Amount.fromUnsafe(1)._unsafeUnwrap(),
 							},