+++ packages/hardware-ledger/test/physical-devices/ledgerNano.test.integration.ts
@@ -3,7 +3,7 @@
  */
 
 /* eslint-disable */
-/*
+
 import { SemVerT, SignTXOutput } from '@radixdlt/hardware-wallet'
 import { log } from '@radixdlt/util'
 import { Subscription } from 'rxjs'
@@ -18,15 +18,16 @@ import {
 } from '@radixdlt/crypto'
 import {
 	BuiltTransactionReadyToSign,
-	NetworkT,
+	Network,
 	uint256FromUnsafe,
 } from '@radixdlt/primitives'
 import { Transaction } from '@radixdlt/tx-parser/dist/transaction'
 import { TransactionT } from '@radixdlt/tx-parser'
 import { AccountAddress } from '@radixdlt/account'
 import { stringifyUInt256 } from '@radixdlt/tx-parser/dist/tokens'
+
 // @ts-ignore
-import TransportNodeHid from '@aleworm/hw-transport-node-hid'
+import TransportNodeHid from '@ledgerhq/hw-transport-node-hid'
 
 describe('hw_ledger_integration', () => {
 	let ledgerNano: LedgerNanoT
@@ -55,7 +56,7 @@ describe('hw_ledger_integration', () => {
 	it('getVersion_integration', async done => {
 		const transport = await TransportNodeHid.create()
 
-		ledgerNano = await LedgerNano.connect(transport, {
+		ledgerNano = await LedgerNano.connect(transport as any, {
 			deviceConnectionTimeout: 10_000,
 		})
 		const hardwareWallet = HardwareWalletLedger.from(ledgerNano)
@@ -78,14 +79,16 @@ describe('hw_ledger_integration', () => {
 	it('getPublicKey_integration', async done => {
 		const transport = await TransportNodeHid.create()
 
-		ledgerNano = await LedgerNano.connect(transport, {
+		ledgerNano = await LedgerNano.connect(transport as any, {
 			deviceConnectionTimeout: 10_000,
 		})
 		const hardwareWallet = HardwareWalletLedger.from(ledgerNano)
 
 		const subs = new Subscription()
 
-		const path = HDPathRadix.fromString(`m/44'/1022'/2'/1/3`)._unsafeUnwrap()
+		const path = HDPathRadix.fromString(
+			`m/44'/1022'/2'/1/3`,
+		)._unsafeUnwrap()
 		const displayAddress = true
 
 		const expectedPubKeyHex =
@@ -99,11 +102,11 @@ describe('hw_ledger_integration', () => {
 			console.log(`🔮 expected path: ${path.toString()}`)
 			const accountAddress = AccountAddress.fromPublicKeyAndNetwork({
 				publicKey: expectedPubKey,
-				network: NetworkT.BETANET,
+				network: Network.MAINNET,
 			})
 			const wrongAccountAddress = AccountAddress.fromPublicKeyAndNetwork({
 				publicKey: expectedPubKey,
-				network: NetworkT.MAINNET,
+				network: Network.MAINNET,
 			})
 			console.log(
 				`🔮 expected address: '${accountAddress.toString()}' ([wrong]mainnet: '${wrongAccountAddress.toString()}')`,
@@ -131,7 +134,7 @@ describe('hw_ledger_integration', () => {
 	it('doKeyExchange_integration', async done => {
 		const transport = await TransportNodeHid.create()
 
-		ledgerNano = await LedgerNano.connect(transport, {
+		ledgerNano = await LedgerNano.connect(transport as any, {
 			deviceConnectionTimeout: 10_000,
 		})
 		const hardwareWallet = HardwareWalletLedger.from(ledgerNano)
@@ -157,7 +160,7 @@ describe('hw_ledger_integration', () => {
 			const accountAddressOfOtherParty = AccountAddress.fromPublicKeyAndNetwork(
 				{
 					publicKey: publicKeyOfOtherParty,
-					network: NetworkT.BETANET,
+					network: Network.MAINNET,
 				},
 			)
 
@@ -195,7 +198,7 @@ describe('hw_ledger_integration', () => {
 		async done => {
 			const transport = await TransportNodeHid.create()
 
-			ledgerNano = await LedgerNano.connect(transport, {
+			ledgerNano = await LedgerNano.connect(transport as any, {
 				deviceConnectionTimeout: 20_000,
 			})
 			const hardwareWallet = HardwareWalletLedger.from(ledgerNano)
@@ -245,7 +248,7 @@ describe('hw_ledger_integration', () => {
 					.doSignTransaction({
 						path,
 						nonXrdHRP: 'btc',
-						tx
+						tx,
 					})
 					.subscribe(
 						(result: SignTXOutput) => {
@@ -277,7 +280,7 @@ describe('hw_ledger_integration', () => {
 	it('doSignHash_integration', async done => {
 		const transport = await TransportNodeHid.create()
 
-		ledgerNano = await LedgerNano.connect(transport, {
+		ledgerNano = await LedgerNano.connect(transport as any, {
 			deviceConnectionTimeout: 10_000,
 		})
 		const hardwareWallet = HardwareWalletLedger.from(ledgerNano)
@@ -288,7 +291,9 @@ describe('hw_ledger_integration', () => {
 			`I'm testing Radix awesome hardware wallet!`,
 		)
 
-		const path = HDPathRadix.fromString(`m/44'/1022'/2'/1/3`)._unsafeUnwrap()
+		const path = HDPathRadix.fromString(
+			`m/44'/1022'/2'/1/3`,
+		)._unsafeUnwrap()
 
 		console.log(`🔮 Path: ${path.toString()}`)
 		console.log(`🔮 Hash: ${hashToSign.toString('hex')}`)
@@ -313,4 +318,3 @@ describe('hw_ledger_integration', () => {
 		)
 	}, 40_000)
 })
-*/